Switch from Command-T to fzf
I didn’t want to do this (I like Command-T and it’s always worked well),
but it started freezing when using Watchman, and then when I disabled
Watchman it instead gave me a list of files that I couldn’t interact
with. I don’t want to debug this right now but maybe one day.